requirements.txt
django_countries==7.0
django-crispy-forms==1.10.0
Django==3.1.5
Pillow==8.1.2
